// src/shared-report-level/enums/window-function.enum.ts
export enum WindowFunctionEnum {
    SumW = 0,
    AvgW = 1,
    CountW = 2,
    MinW = 3,
    MaxW = 4,
    VarW = 5,
    FirstW = 6,
    LastW = 7,
    SumRT = 8,
    AvgRT = 9,
    CountRT = 10,
    MinRT = 11,
    MaxRT = 12,
    VarRT = 13,
    Rank = 14,
    DenseRank = 15,
    RowNumber = 16,
    Lead = 17,
    Lag = 18,
    NTile = 19
}